+++ core/src/main/java/de/timuuuu/moneymaker/activities/popup/MuteActivity.java
@@ -1,7 +1,7 @@
 package de.timuuuu.moneymaker.activities.popup;
 
 import de.timuuuu.moneymaker.MoneyMakerAddon;
-import de.timuuuu.moneymaker.chat.MoneyChatMessage;
+import de.timuuuu.moneymaker.moneychat.util.MoneyChatMessage;
 import de.timuuuu.moneymaker.moneychat.protocol.packets.PacketUserMute;
 import de.timuuuu.moneymaker.utils.Util;
 import net.labymod.api.Laby;
