+++ core/src/main/java/de/timuuuu/moneymaker/events/MoneyChatReceiveEvent.java
@@ -1,6 +1,6 @@
 package de.timuuuu.moneymaker.events;
 
-import de.timuuuu.moneymaker.chat.MoneyChatMessage;
+import de.timuuuu.moneymaker.moneychat.util.MoneyChatMessage;
 import net.labymod.api.event.Event;
 
 public class MoneyChatReceiveEvent implements Event {
