+++ core/src/main/java/de/timuuuu/moneymaker/listener/EntityRenderListener.java
@@ -1,7 +1,7 @@
 package de.timuuuu.moneymaker.listener;
 
 import de.timuuuu.moneymaker.MoneyMakerAddon;
-import de.timuuuu.moneymaker.chat.ChatUtil;
+import de.timuuuu.moneymaker.utils.ChatUtil;
 import de.timuuuu.moneymaker.event.ArmorStandRenderEvent;
 import de.timuuuu.moneymaker.settings.AddonSettings;
 import de.timuuuu.moneymaker.utils.Util;
